Move graphql_to_quick_reply into _quick_reply.py
